// src/shared/models/member.model.ts
export interface Member {
        linkedTo: string;
        memberImage: string;
        updatedOn: Date;
        _id: string;
        createDate: Date;
        cycleStartDate: any;
        email: string;
        firstName: string;
        address: string;
        cycleEndDate: any;
        lastName: string;
        parent: string;
        phone: string;
    }